Log per-minute average humidity/temperature to EEPROM.
Store humidity/temperature data in EEPROM, to enable data logging without
having something connected to the serial port.

On (the next) startup, dump all the samples stored in the EEPROM to the
serial port before resuming measurements. This allows retrieval of samples
from the last data logging session.

Since the EEPROM is limited in size (1024 bytes), we must be minimize the
size of the stored samples in order to get a useful session length (hours).
Therefore, we don't store each sample, but instead only store the per-minute
averages. Also, instead of storing the entire 4-byte float values, compress
them into a single byte, by multiplying the float by 4, and truncating it
into a single byte. This gives a theoretical range of values from 0 to
63.75 in steps of 0.25. (We thus assume that the humidity/temperature will
never exceed 63.74 %/C). Thus we end up writing 2 bytes per minute to the
EEPROM (1 byte each for the average humidity/temperature). This gives a
total session length of 512 minutes (8 hours and 32 minutes).

The samples are written to consecutive EEPROM addresses, wrapping around when
hitting the end of the EEPROM. There is currently no mechanism to store the
last written EEPROM address. Thus, if we wrap around, there is no visible
boundary between most recent and least recent sample. Furthermore, before we
wrap around, we have no way to detect the end of the last run, if it was
preceded by a longer run (that wrote further into the EEPROM).