A simple sketch for sequencing power-up of external equipment.
Situation

There are multiple pieces of equipment that should be powered on more or less
simultaneously. For example a collection of audio/video equipment including a
games console, a TV, a pre-amp, and several power-amp modules.

Problem

Several of these components consume a significant inrush current for a short
time period when powered up. Each inrush current spike is not problematic, but
when all these devices are powered on simultaneously, the resulting combined
inrush current spike is large enough to trigger the circuit breaker, and leave
everything dark.

Solution

Separate the inrush current spikes by powering up each component in sequence.
Use a half-second delay between each component to allow the inrush current for
a component to (mostly) subside before starting the next component. Implement
this sequencing by controlling the power supply to each component with a
relay, and then control all of these relays from an Arduino.

The Arduino code to control this is very simple: On power-up, make all pins
output low, and then turn each output pin high in sequence, with a half-second
delay between pins.