Start implementing a low-power idle mode
This puts the device into a low-power sleep when IDLE is returned from
process_inputs(). The sleep mode is only entered when there is no input
to be processed, and the RCN network is also idle. When entering the
low-power sleep, we first blink the red LED three times, to tell the user
that we're going to sleep. We then disable the PWM outputs, and call
Sleepy::powerDown() to enter the low-power sleep mode.

Running this sketch on a 3.3V JeeNode v6 yields a current of ~0.54mA in
the sleep mode (~= 1.8mW)